Alter description
First, refer to the language dialect the same as in our READMEs.

Second, remove the Haskell reference.  I will shortly add Rust and so it
would be unfair to mention only Haskell.

Third, remove the disclaimer that the compiler is at an early stage.
Partly this is because I intend to improve the compiler.  Also, the
disclaimer is not that important because people are unlikely to see the
website before the compiler is more impressive.